Trying to add win support
This is from 2016, forgot to publish

+++ pyautobrightness/pyautobrightness.py
@@ -1,53 +1,60 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 # PIL, Pygame, os(xbacklight), numpy
 
-import os,sys,configparser,subprocess,pygame,pygame.camera,numpy,time
-from PIL import Image
+#ARRUMAR ONDE TIRA FOTO E CONVERTE DEPOIS ARRUMAR O PRIMEIRA CONFIGURAÇÃO
 
+import os,sys,subprocess,numpy,time,ConfigParser
+
+if os.name == "nt":
+	sClean = "cls"
+else:
+	sClean = "clear"
+	import pygame, pygame.camera #don't think I need this
+
+Configure = ConfigParser.ConfigParser()
+
+ConfigDirRoot = "%s/.config/" %(os.path.expanduser('~'))
 ConfigDir="%s/.config/autobrighness" %(os.path.expanduser('~'))
 ConfigFile="%s/config" %ConfigDir
 Save="%s/.current.jpg" %ConfigDir
 	
-Configure = configparser.ConfigParser()
 Configure.read(ConfigFile)
 
 try:
-	X=Configure["DoNotChange"]["X"]
-	Y=Configure["DoNotChange"]["Y"]
-	TransitionTime=int(Configure["CONFIG"]["Transition"])
-	MODE=Configure["CONFIG"]["Mode"]
-	NumberOfPictures=int(Configure["CONFIG"]["NumberOfPictures"])
-	TimeBetweenPictures=int(Configure["CONFIG"]["TimeBetweenPictures"])
-	Deviece=Configure["CONFIG"]["Deviece"]
+	X=Configure.get("DoNotChange","X")
+	Y=Configure.get("DoNotChange","Y")
+	if os.name != "nt":
+		TransitionTime=int(Configure.get("CONFIG","Transition"))
+	MODE=Configure.get("CONFIG","Mode")
+	NumberOfPictures=int(Configure.get("CONFIG","NumberOfPictures"))
+	TimeBetweenPictures=int(Configure.get("CONFIG","TimeBetweenPictures"))
+	Deviece=Configure.get("CONFIG","Deviece")
 except:
 	pass
 
 try:
-	Brilho_MIN=int(Configure["CONFIG"]["Minimum_Brightness"])
-	Brilho_MAX=int(Configure["CONFIG"]["Maximum_Brightness"])
+	Brilho_MIN=int(Configure.get("CONFIG","Minimum_Brightness"))
+	Brilho_MAX=int(Configure.get("CONFIG","Maximum_Brightness"))
 except:
 	pass
-
 
 ##from config
 
 def TestIfConfExist():
-	try:
-		os.stat(ConfigDir)
-		try:
-			os.stat(ConfigFile)
-		except:
-			return(False)
-		return(True)
-	except:
-		os.mkdir(ConfigDir) 
+	if os.path.isdir(ConfigDirRoot) is False:
+		os.mkdir(ConfigDirRoot)
+	if os.path.isdir(ConfigDir) is False:
+		os.mkdir(ConfigDir)
 		return(False)
+	if os.path.isfile(ConfigFile) is False:
+		return(False)
+	return(True)
 
 def TestIfMaxMinExist():
 	try:
-		Brilho_MIN=int(Configure["CONFIG"]["Minimum_Brightness"])
-		Brilho_MAX=int(Configure["CONFIG"]["Maximum_Brightness"])
+		Brilho_MIN=int(Configure.get("CONFIG","Minimum_Brightness"))
+		Brilho_MAX=int(Configure.get("CONFIG","Maximum_Brightness"))
 		return(True)
 	except:
 		return(False)
@@ -56,14 +63,18 @@ def CreateConfigFile():
 		ConfCreate = open(ConfigFile, 'w')
 		ConfCreate.write("[CONFIG]")
 		ConfCreate.write("\n")
-		ConfCreate.write("Deviece = /dev/video0")
+		if os.name == "nt":
+			ConfCreate.write("Deviece = 0")
+		else:
+			ConfCreate.write("Deviece = /dev/video0")
 		ConfCreate.write("\n")
 		ConfCreate.write("Minimum_Brightness =")
 		ConfCreate.write("\n")
 		ConfCreate.write("Maximum_Brightness =")
 		ConfCreate.write("\n")
-		ConfCreate.write("Transition = 4000")
-		ConfCreate.write("\n")
+		if os.name != "nt":
+			ConfCreate.write("Transition = 4000")
+			ConfCreate.write("\n")
 		ConfCreate.write("numberofpictures = 2")
 		ConfCreate.write("\n")
 		ConfCreate.write("timebetweenpictures = 1")
@@ -78,53 +89,101 @@ def CreateConfigFile():
 		ConfCreate.write("y =")
 		ConfCreate.write("\n")
 		ConfCreate.close()
+		print("Configuration created, run the program again to calibrate.")
+		exit(0)
 
-def getch():
-	import tty, termios
-	fd = sys.stdin.fileno()
-	old_settings = termios.tcgetattr(fd)
-	try:
-		tty.setraw(fd)
-		ch = sys.stdin.read(1)
-	finally:
-		termios.tcsetattr(fd, termios.TCSADRAIN, old_settings)
-	return ch
+try:
+	from msvcrt import getch
+except ImportError:
+	def getch():
+		import tty, termios
+		fd = sys.stdin.fileno()
+		old_settings = termios.tcgetattr(fd)
+		try:
+			tty.setraw(fd)
+			ch = sys.stdin.read(1)
+		finally:
+			termios.tcsetattr(fd, termios.TCSADRAIN, old_settings)
+		return ch
 
-def GetBRInput( Txt ):
-	var=float(subprocess.check_output("xbacklight -get", shell=True))
+def fWinReturnScreenBrightness():
+	import wmi
+	oC = wmi.WMI(namespace='wmi')
+	oScreen = oC.WmiMonitorBrightness()[0]
+	oScreen
+	return(float(oScreen.CurrentBrightness))
+
+def fWinIncDecScreenBright(A, Mod):
+	if Mod == "Set":
+		import wmi
+		oC = wmi.WMI(namespace='wmi')
+		oScreen = oC.WmiMonitorBrightnessMethods()[0]
+		oScreen.WmiSetBrightness(A, 0)
+	elif Mod == "Sub":
+		import wmi
+		oC = wmi.WMI(namespace='wmi')
+		oScreen = oC.WmiMonitorBrightnessMethods()[0]
+		fNewBright = fWinReturnScreenBrightness()+A
+		if fNewBright >= 0 and fNewBright <=100:
+			oScreen.WmiSetBrightness(fNewBright, 0)
+		
+def iRoundNumber(Num):
+	Dec=int(Num/10)*10
+	Unit = Num-Dec
+	if Unit >= 0 and Unit <= 3:
+		return(Dec+0)
+	elif Unit > 3 and Unit < 7:
+		return(Dec+5)
+	else:
+		return(Dec+10)
+
+def sGetBrightnessInput( Txt ):
+	if os.name == "nt":
+		fVar=fWinReturnScreenBrightness()
+	else:
+		fVar=float(subprocess.check_output("xbacklight -get", shell=True))
 	while True:
-		subprocess.call("clear")
+		os.system(sClean)
 		print(Txt)
-		Now="#"*int(var)
-		Then="-"*(100-int(var))
-		Perc="%.1f" %var
+		Now="#"*int(fVar)
+		Then="-"*(100-int(fVar))
+		Perc="%.1f" %fVar
 		print("["+Now+Then+"] ["+Perc+"%]")
 		
 		keypress=getch()
-		if keypress == "+":
-			subprocess.call("xbacklight -time 0 -inc 1", shell=True)
-			var=float(subprocess.check_output("xbacklight -get", shell=True))
-		elif keypress == "-":
-			subprocess.call("xbacklight -time 0 -dec 1", shell=True)
-			var=float(subprocess.check_output("xbacklight -get", shell=True))
-		elif keypress == "y":
-			return(var)
-		elif keypress == "n" or keypress == "q":
-			subprocess.call("clear")
+
+		if keypress == "+" or ord(keypress) == 43:
+			if os.name == "nt":
+				fWinIncDecScreenBright(10, "Sub")
+				fVar = fWinReturnScreenBrightness()
+			else:
+				subprocess.call("xbacklight -time 0 -inc 1", shell=True)
+				fVar=float(subprocess.check_output("xbacklight -get", shell=True))
+		elif keypress == "-" or ord(keypress) == 45:
+			if os.name == "nt":
+				fWinIncDecScreenBright(-10, "Sub")
+				fVar = fWinReturnScreenBrightness()
+			else:
+				subprocess.call("xbacklight -time 0 -dec 1", shell=True)
+				fVar=float(subprocess.check_output("xbacklight -get", shell=True))
+		elif keypress == "y" or ord(keypress) == 121:
+			return(fVar)
+		elif keypress == "n" or keypress == "q" or ord(keypress) == 110 or ord(keypress) == 113 or ord(keypress) == 27:
+			os.system(sClean)
 			print("Exiting without save...")
 			exit(0)
 		else:
 			pass
 
 def CreateMinMax():
-	Configure.read(ConfigFile)
-	Brilho_MIN=int(GetBRInput("Select the MINIMUM brightness PyAutoBrightness should use (use + and - keys, then press y to confirm or n to exit)"))
+	# Configure.read(ConfigFile)
+	Brilho_MIN=int(sGetBrightnessInput("Select the MINIMUM brightness PyAutoBrightness should use (use + and - keys, then press y to confirm or n to exit)"))
 	Configure.set("CONFIG","Minimum_Brightness",str(Brilho_MIN))
-	subprocess.call("clear")
+	os.system(sClean)
 	print("Saving Minimum Brightness...")
 	time.sleep(1)
-	subprocess.call("clear")
-	Brilho_MAX=int(GetBRInput("Select the MAXIMUM brightness PyAutoBrightness should use (use + and - keys, then press y to confirm or n to exit)"))
+	os.system(sClean)
+	Brilho_MAX=int(sGetBrightnessInput("Select the MAXIMUM brightness PyAutoBrightness should use (use + and - keys, then press y to confirm or n to exit)"))
 	Configure.set("CONFIG","Maximum_Brightness",str(Brilho_MAX))
 	with open(ConfigFile, 'w') as configfile:
 			Configure.write(configfile)
@@ -132,19 +191,41 @@ def CreateMinMax():
 	time.sleep(1)
 	
 def CaptureImage( im_file ): #local de salvamento, retorna porcentagem de brilho
-	pygame.init()
-	pygame.camera.init()
-	cam = pygame.camera.Camera(Deviece, (640,480))
-	cam.start()
-	image = cam.get_image()
-	cam.stop()
-	pygame.image.save(image, im_file)
-	from PIL import ImageStat
-	im = Image.open(im_file).convert('L')
-	stat = ImageStat.Stat(im)
-	perc=int(stat.mean[0]*100/255)
-	os.remove(im_file)
-	return(perc)
+	# try:
+	# 	Deviece=Configure.get("CONFIG","Deviece")
+	# except:
+	# 	pass
+	if os.name == "nt":
+		import cv2
+		from PIL import Image
+		from PIL import ImageStat
+		#take the pic from webcam
+		cam = cv2.VideoCapture(int(Deviece))
+		_,image = cam.read()
+		image = cv2.cvtColor(image,cv2.COLOR_BGR2RGB)
+		image = Image.fromarray(image)
+		#get the brightness
+		stat = ImageStat.Stat(image)
+		perc = int(stat.mean[0]*100/255)
+		cam.release()
+		perc = iRoundNumber(perc)
+		return(perc)
+	else:
+		import pygame
+		pygame.init()
+		pygame.camera.init()
+		cam = pygame.camera.Camera(Deviece, (640,480))
+		cam.start()
+		image = cam.get_image()
+		cam.stop()
+		pygame.image.save(image, im_file)
+		from PIL import ImageStat
+		im = Image.open(im_file).convert('L')
+		stat = ImageStat.Stat(im)
+		perc=int(stat.mean[0]*100/255)
+		os.remove(im_file)
+		perc = iRoundNumber(perc)
+		return(perc)
 	
 def Captures( num_pic, Interval, Mode ):
 	if num_pic == 1:
@@ -175,16 +256,16 @@ def Captures( num_pic, Interval, Mode ):
 		else:
 			return(numpy.median(numpy.array(value)))
 	else:
-		print("Must take at least one picture, check the config file")
+		print("Must take at least one picture, check the configuration file.")
 		exit(1)
 	
 def RunFirstCalibration():
-	subprocess.call("clear")
-	Deviece=Configure["CONFIG"]["Deviece"]
+	os.system(sClean)
+	# Deviece=Configure.get("CONFIG","Deviece")
 	print("Cover your webcam and press y, so PyAutoBrightness can see the minimum brightness value your webcam can capture and add it to the equation")
 	while True:
 		test=getch()
-		if test == "Y" or test == "y":
+		if test == "Y" or test == "y" or ord(test) == 121 or test == 89:
 			break
 	MinX=CaptureImage(Save)
 	try:
@@ -192,11 +273,12 @@ def RunFirstCalibration():
 		MinY=Brilho_MIN
 	except:
 		Configure.read(ConfigFile)
-		MinY=int(Configure["CONFIG"]["Minimum_Brightness"])
-	time.sleep(2)
+		MinY=float(Configure.get("CONFIG","Minimum_Brightness"))
+	time.sleep(1)
 	print("You can uncover it now!")
-	MaxY=GetBRInput("Now go to the brightest place you can, set the screen brightness as you please and press Y (don't worry if you can't reach the brightest place you use your PC, you can calibrate PyAutoBrightness after for other values:")
+	MaxY=sGetBrightnessInput("Now go to the brightest place you can, set the screen brightness as you please and press Y (don't worry if you can't reach the brightest place you use your PC, you can calibrate PyAutoBrightness after for other values:")
 	MaxX= Captures( 3, 1, "precise")
+	print(MaxX,MaxY)
 	X="%s,%s" %(str(int(round(MinX,0))), str(int(round(MaxX,0))))
 	print(X)
 	Y="%s,%s" %(str(int(round(MinY,0))), str(int(round(MaxY,0))))
@@ -204,7 +286,7 @@ def RunFirstCalibration():
 	Configure.set("DoNotChange","y", Y)
 	with open(ConfigFile, 'w') as configfile:
 			Configure.write(configfile)
-	print("PyAutoBrightness sucessful calibrated! Try run pyautobrightness -c in various light conditions; the more you run it, the better PyAutoBrightness will get! Press any key to exit.")
+	print("PyAutoBrightness sucessful calibrated! Try run pyautobrightness -c in Various light conditions; the more you run it, the better PyAutoBrightness will get! Press any key to exit.")
 	getch()
 
 def ConvertIntoList( str ):
@@ -212,67 +294,84 @@ def ConvertIntoList( str ):
 	LSTFloat= []
 	for i in LST:
 		try:
-			LSTFloat.append(float(i.strip()))
+			LSTFloat.append(int(i.strip()))
 		except ValueError:
 			print("""ERROR: no numbers in X Y, try to calibrate again.""")
 			exit(1)
 	return(LSTFloat)
 
-#calibrate
-
 def AddtoEquation( new_coord, list_coord ):
-	x= int(new_coord[0])
-	y= int(new_coord[1])
+	"""this function takes new coordinates (x,y), and fit them inside the already current
+	coordinates, that are saved in the config file. 
+	RETURNS A LIST OF LISTS"""
 	ListX = list_coord[0]
 	ListY = list_coord[1]
+
+	for i in range(0, len(ListX)):
+		if ListX[i]%5 != 0:
+			ListX[i] = iRoundNumber(ListX[i])
+			
+	x= iRoundNumber(float(new_coord[0]))
+	y= float(new_coord[1])
+
 	if x not in ListX:
-		ListX.append(float(x))
+		ListX.append(int(x))
 		ListX.sort()
-		SlotY= ListX.index(x)
-		ListY.insert(SlotY, float(y))
+		SlotY = ListX.index(x)
+		ListY.insert(SlotY, int(y))
 	else:
-		SlotY= ListX.index(x)
-		OldY=ListY[SlotY]
-		NewY=(OldY+y)/2
-		ListY.pop(SlotY)
-		ListY.insert(SlotY, float(NewY))
-	return([ListX, ListY])	
+		SlotY = ListX.index(x)
+		OldY = ListY[SlotY]
+		NewY = float((OldY+y)/2)
+		ListY[SlotY] = int(NewY)
+	return([ListX, ListY])
 
 def Calibrate():
 	XList = ConvertIntoList(X)
 	YList = ConvertIntoList(Y)
-	var=float(subprocess.check_output("xbacklight -get", shell=True))
+	if os.name == "nt":
+		fVar=fWinReturnScreenBrightness()
+	else:
+		fVar=float(subprocess.check_output("xbacklight -get", shell=True))
 	while True:
-		subprocess.call("clear")
+		os.system(sClean)
 		print("Select a good brightness for the current light in your room (use + and - keys, then press y to confirm or n to exit)")
-		Now="#"*int(var)
-		Then="-"*(100-int(var))
-		Perc="%.1f" %var
+		Now="#"*int(fVar)
+		Then="-"*(100-int(fVar))
+		Perc="%.1f" %fVar
 		print("["+Now+Then+"] ["+Perc+"%]")
 		keypress=getch()
-		if keypress == "+":
-			subprocess.call("xbacklight -time 0 -inc 1", shell=True)
-			var=float(subprocess.check_output("xbacklight -get", shell=True))
-		elif keypress == "-":
-			subprocess.call("xbacklight -time 0 -dec 1", shell=True)
-			var=float(subprocess.check_output("xbacklight -get", shell=True))
-		elif keypress == "y":
-			Ycurrent=int(var)
+		if keypress == "+" or ord(keypress) == 43:
+			if os.name == "nt":
+				fWinIncDecScreenBright(10, "Sub")
+				fVar = fWinReturnScreenBrightness()
+			else:
+				subprocess.call("xbacklight -time 0 -inc 1", shell=True)
+				fVar=float(subprocess.check_output("xbacklight -get", shell=True))
+		elif keypress == "-" or ord(keypress) == 45:
+			if os.name == "nt":
+				fWinIncDecScreenBright(-10, "Sub")
+				fVar = fWinReturnScreenBrightness()
+			else:
+				subprocess.call("xbacklight -time 0 -dec 1", shell=True)
+				fVar=float(subprocess.check_output("xbacklight -get", shell=True))
+		elif keypress == "y" or ord(keypress) == 121:
+			Ycurrent=int(fVar)
 			break
-		elif keypress == "n" or keypress == "q":
-			subprocess.call("clear")
+		elif keypress == "n" or keypress == "q" or ord(keypress) == 110 or ord(keypress) == 113:
+			os.system(sClean)
 			print("Exiting without save...")
 			exit(0)
 		else:
 			pass
 	Xcurrent=Captures(3, 1, "precise")
 	ListSAVE= AddtoEquation( (Xcurrent,Ycurrent), (XList, YList))
-	def SaveList( List, destvar, destfile ):
+	def SaveList( List, destfVar, destfile ):
 		File=""
 		for i in List:
 			File+=",%s" %int(i)
 		print(File[1:])
-		Configure.set("DoNotChange",destvar,File[1:])
+		Configure.set("DoNotChange",destfVar,File[1:])
 		with open(destfile, 'w') as configfile:
 			Configure.write(configfile)			
 	SaveList(ListSAVE[0], "x", ConfigFile)
@@ -289,4 +388,8 @@ def RUN():
 		BrightValue=Brilho_MAX
 	elif BrightValue < Brilho_MIN:
 		BrightValue=Brilho_MIN
-	subprocess.call("xbacklight -time %s -steps %s -set %i" %(TransitionTime,TransitionTime,int(round(BrightValue,0))), shell=True)
+	print(BrightValue)
+	if os.name == "nt":
+		fWinIncDecScreenBright(int(round(BrightValue,0)), "Set")
+	else:
+		subprocess.call("xbacklight -time %s -steps %s -set %i" %(TransitionTime,TransitionTime,int(round(BrightValue,0))), shell=True)
